# api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Nova RMS API
  version: 1.0.0
  description: REST surface for Identity, Tenant Registry, RBAC, Module Registry, Superadmin Ops, Billing, Support, POS, and Inventory modules.
servers:
  - url: https://api.nova-rms.com/v1
    description: Production
  - url: https://staging-api.nova-rms.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /superadmin/registrations:
    get:
      tags: [Superadmin]
      summary: List registration requests
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        '200':
          description: Registrations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Registration'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /superadmin/registrations/{registrationId}/modules:
    patch:
      tags: [Superadmin]
      summary: Update module toggles for a pending registration
      parameters:
        - $ref: '#/components/parameters/RegistrationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modules]
              properties:
                modules:
                  type: array
                  items:
                    $ref: '#/components/schemas/RegistrationModuleToggle'
      responses:
        '200':
          description: Updated module toggles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      modules:
                        type: array
                        items:
                          $ref: '#/components/schemas/RegistrationModuleToggle'
        '400':
          $ref: '#/components/responses/ValidationError'
  /superadmin/registrations/{registrationId}/decision:
    post:
      tags: [Superadmin]
      summary: Approve or reject registration
      parameters:
        - $ref: '#/components/parameters/RegistrationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
                plan_id:
                  type: string
                  format: uuid
                modules:
                  type: array
                  items:
                    $ref: '#/components/schemas/RegistrationModuleToggle'
                reason:
                  type: string
      responses:
        '200':
          description: Decision stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationStatus'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    RegistrationId:
      name: registrationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    RegistrationStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, approved, rejected]
        submitted_at:
          type: string
          format: date-time
        decided_at:
          type: string
          format: date-time
        reviewer_id:
          type: string
          format: uuid
    Registration:
      allOf:
        - $ref: '#/components/schemas/RegistrationStatus'
        - type: object
          properties:
            business:
              $ref: '#/components/schemas/RegistrationInput/properties/business'
            owner:
              $ref: '#/components/schemas/RegistrationInput/properties/owner'
            tenant_id:
              type: string
              format: uuid
              nullable: true
            modules:
              type: array
              items:
                $ref: '#/components/schemas/RegistrationModuleToggle'
    RegistrationModuleToggle:
      type: object
      required: [key, name, enabled]
      properties:
        key:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        category:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION, AUTHN, AUTHZ, NOT_FOUND, CONFLICT, RATE_LIMIT, INTERNAL]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    ValidationError:
      description: Request invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ConflictError:
      description: Operation conflicts with state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'